Fixed: Event files

+++ actions/event/deletefile.php
@@ -1,46 +1,55 @@
 <?php 
 
-$guid = get_input('guid');
-$filename = get_input('file');
+	$guid = get_input('guid');
+	$filename = get_input('file');
+	$time = get_input('t');
 
-if(!empty($guid) && !empty($filename))
-{
-	if($entity = get_entity($guid))
+	if(!empty($guid) && !empty($filename))
 	{
-		if($entity->getSubtype() == Event::SUBTYPE)
+		if($entity = get_entity($guid))
 		{
-			$event = $entity;
-		}
-	}
-	
-	if($event->canEdit())
-	{
-		$files = json_decode($event->files);
-		$i = 0;
-		foreach($files as $file)
-		{
-			if(strtolower($file->file) == strtolower($filename))
+			if($entity->getSubtype() == Event::SUBTYPE)
 			{
-				header('Content-Type: '.$file->mime);
-				header('Content-Disposition: Attachment; filename='.$file->file);
-				header('Pragma: no-cache');
-				
-				$prefix = "events/".$event->getGUID()."/files/";
-				
-				$fileHandler = new ElggFile();
-				$fileHandler->setFilename($prefix . $file->file);
-				if($fileHandler->owner_guid == $event->owner_guid)
-				{
-					$fileHandler->delete();
-					unset($files[$i]);
-				}
+				$event = $entity;
 			}
-			$i++;
 		}
-		$event->files = json_encode($files);
+
+		if($event->canEdit())
+		{
+			$files = json_decode($event->files, true);
+			$i = 0;
+			
+			foreach($files as $file)
+			{
+				if((strtolower($file['file']) == strtolower($filename)) && ($time == $file['time_uploaded']))
+				{
+					$prefix = "events/".$event->getGUID()."/files/";
+
+					$fileHandler = new ElggFile();
+					$fileHandler->setFilename($prefix . $file['file']);
+					$fileHandler->owner_guid = $event->owner_guid;
+					
+					$fileHandler->delete();
+					
+					if(count($files) == 1)
+					{
+						$files = array();
+					}
+					else
+					{
+						unset($files[$i]);
+					}
+					break;
+				}
+
+				$i++;
+			}
+			
+			$event->files = json_encode($files);
+		}
+		
+		forward(REFERER);
 	}
-	forward(REFERER);
-}
 	
-register_error(elgg_echo('event_manager:event:file:notfound:text'));
-forward(REFERER);
+	register_error(elgg_echo('event_manager:event:file:notfound:text'));
+	forward(REFERER);
